remove-requirements.txt
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-ipware==4.0.2
idna==3.3
oauthlib==3.2.0
Pillow==9.0.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
urllib3==1.26.8